==> 3a) MSc Project 2019 Aaron Dougherty Word2Vec Models.ipynb <==
#Initial Downloads and imports

import tensorflow as tf
from tensorflow import keras
print(tf.__version__)

import pandas as pd
import numpy as np
import re
import matplotlib.pyplot as plt

from datetime import datetime
from dateutil.parser import parse

from sklearn import svm
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn import metrics
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix
from sklearn.model_selection import train_test_split
from sklearn.utils import shuffle, resample

import gensim
import warnings
warnings.filterwarnings(action='ignore', category=UserWarning, module='gensim')
warnings.simplefilter(action='ignore', category=FutureWarning)
from gensim.models.keyedvectors import KeyedVectors
from gensim.models import Word2Vec
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from gensim.models.phrases import Phrases, Phraser

import matplotlib.pyplot as plt

import nltk
nltk.download('punkt')
from nltk import word_tokenize

import sys
import csv
csv.field_size_limit(sys.maxsize)

from keras.preprocessing import sequence
from keras.utils import np_utils
from keras.models import Model
from keras.layers import Dense, Dropout, Activation, Lambda, Input, concatenate
from keras.layers.embeddings import Embedding
from keras.layers.recurrent import LSTM, SimpleRNN, GRU
from keras.preprocessing.text import Tokenizer
from collections import defaultdict
from keras.layers.convolutional import Convolution1D

# Mount drive to import necessary files:
from google.colab import drive
drive.mount('/content/drive')

# def prepare_text_data(titles, content):
#   """ Method to load in data and prepare final pandas.DataFrame for inputting
#       machine learning algorithms

#   Parameters
#   ----------
#   titles : str
#       File name of numpy array of title document embedding vectors
#   content : str
#       File name of numpy array of content document embedding vectors
#   """
#   # file path for main csv file of all text data
#   path = '/content/drive/My Drive/Colab Notebooks/csvData/title_content.csv'
#   # read in main csv file into pandas.DataFrame
#   df = pd.read_csv(path, encoding='ISO-8859-1', engine='python', error_bad_lines=False)
#   # file path for sentiment and finance data
#   path2 = '/content/drive/My Drive/Colab Notebooks/csvData/sent_fin.csv'
#   # read in csv fiel into pandas.DataFrame
#   sent_fin = pd.read_csv(path2, encoding='ISO-8859-1', engine='python', error_bad_lines=False)
#   # remove useless columns
#   sent_fin.drop('Unnamed: 0', axis=1, inplace=True)
#   df.drop('Unnamed: 0', axis=1, inplace=True)
#   print('initial dataframes loaded')
#   # load in document vectors
#   titleVecs = np.load("/content/drive/My Drive/Colab Notebooks/Word2VecData/" + titles,
#                       allow_pickle=True)
#   contentVecs = np.load("/content/drive/My Drive/Colab Notebooks/Word2VecData/" + content,
#                       allow_pickle=True)
#   print('word embeddings loaded')
#   # load document vectors into data frame
#   df['title_vectors'] = pd.Series(titleVecs.tolist())
#   df['content_vectors'] = pd.Series(contentVecs.tolist())
#   print('word embeddings added to dataframe')
  
#   for index, row in df.iterrows(): # convert vectors to numpy arrays
#     row['title_vectors'] = np.array(row['title_vectors'])
#     row['content_vectors'] = np.array(row['content_vectors'])
#   print('text converted to numpy arrays')
#   # find mean document vector per date
#   result_df = df.groupby('date')['title_vectors'].apply(np.mean).to_frame('Title')
#   result_df2 = df.groupby('date')['content_vectors'].apply(np.mean).to_frame('Content')
#   print('mean of document vectors for each date found')
#   # recombine data frames created above
#   result_df['Content'] = pd.Series(result_df2['Content'])
#   result_df = result_df.reset_index().rename(columns={"date" : "Date"})
#   # merge text data with finance and sentiment data
#   result_df = result_df.merge(sent_fin, how='inner', on='Date')
#   return result_df

def prepare_text_data(titles):
  """ Method to load in data and prepare final pandas.DataFrame for inputting
      machine learning algorithms

  Parameters
  ----------
  titles : str
      File name of numpy array of title document embedding vectors
  content : str
      File name of numpy array of content document embedding vectors
  """
  # file path for main csv file of all text data
  path = '/content/drive/My Drive/Colab Notebooks/MSc Project 2019 Aaron Dougherty csvData/title_content.csv'
  # read in main csv file into pandas.DataFrame
  df = pd.read_csv(path, encoding='ISO-8859-1', engine='python', error_bad_lines=False)
  # file path for sentiment and finance data
  path2 = '/content/drive/My Drive/Colab Notebooks/MSc Project 2019 Aaron Dougherty csvData/sent_fin.csv'
  # read in csv fiel into pandas.DataFrame
  sent_fin = pd.read_csv(path2, encoding='ISO-8859-1', engine='python', error_bad_lines=False)
  # remove useless columns
  sent_fin.drop('Unnamed: 0', axis=1, inplace=True)
  df.drop('content', axis=1, inplace=True)
  df.drop('Unnamed: 0', axis=1, inplace=True)
  print('initial dataframes loaded')
  # load in document vectors
  titleVecs = np.load("/content/drive/My Drive/Colab Notebooks/MSc Project 2019 Aaron Dougherty Word2VecData/" + titles,
                      allow_pickle=True)
  print('word embeddings loaded')
  # load document vectors into data frame
  df['title_vectors'] = pd.Series(titleVecs.tolist())
  print('word embeddings added to dataframe')
  
  for index, row in df.iterrows(): # convert vectors to numpy arrays
    row['title_vectors'] = np.array(row['title_vectors'])
  print('text converted to numpy arrays')
  # find mean document vector per date
  result_df = df.groupby('date')['title_vectors'].apply(np.mean).to_frame('Title')
  print('mean of document vectors for each date found')
  # recombine data frames created above
  result_df = result_df.reset_index().rename(columns={"date" : "Date"})
  # merge text data with finance and sentiment data
  result_df = result_df.merge(sent_fin, how='inner', on='Date')
  return result_df

# def explode_df(df):
#   """ Method to explode document vectors into on column per value for feeding
#       into machine learning algorithms

#   Parameters
#   ----------
#   df : pandas.DataFrame
#       data frame to expand
#   """
#   df1 = df.iloc[:, 0:3]
#   df2 = df.iloc[:, 3:]
#   df2['Date'] = df1['Date']
#   splitTitle = pd.DataFrame(df1['Title'].tolist())
#   splitContent = pd.DataFrame(df1['Content'].tolist())
#   splitTitle['Date'] = pd.Series(df1['Date'])
#   splitContent['Date'] = pd.Series(df1['Date'])
#   final = pd.merge(splitTitle, splitContent, how='inner', on='Date')
#   final = final.merge(df2, how='inner', on='Date')
#   return final

def explode_df(df):
  """ Method to explode document vectors into on column per value for feeding
      into machine learning algorithms

  Parameters
  ----------
  df : pandas.DataFrame
      data frame to expand
  """
  df1 = df.iloc[:, 0:2]
  df2 = df.iloc[:, 2:]
  df2['Date'] = df1['Date']
  splitTitle = pd.DataFrame(df1['Title'].tolist())
  splitTitle['Date'] = pd.Series(df1['Date'])
  final = splitTitle.merge(df2, how='inner', on='Date')
  return final

# prepare final data frame
df = prepare_text_data('wordvec2_titles_doc3.npy')

df = explode_df(df)

#Training/Development/Testing Split

def shuffle_df(df):
  """ Method to shuffle rows of a data frame

  Parameters
  ----------
  df : pandas.DataFrame
      Data frame to shuffle
  """
  index = df.index # record data frame indexes
  df = shuffle(df) # ScikitLearn's shuffle method
  df.index = index # reset indexes
  return df

# split into X and y
X = df.iloc[:, 0:-15] # set to -8 to remove financial data and -1 to include; -15 removes sentiment
X.drop('Date', axis=1, inplace=True)
y = df[['Date', 'Direction']]
# split into training, development and testing sets
X_train, X_other, y_train, y_other = train_test_split(X, y, test_size=0.2, shuffle=False)
X_dev, X_test, y_dev, y_test = train_test_split(X_other, y_other, test_size=0.5, shuffle=False)
# remove useless columns
y_dev.drop('Date', axis=1, inplace=True)
y_test.drop('Date', axis=1, inplace=True)

##Dealing with class imbalance:

def recombine(X, y):
  """ Method to recombine the X and y

  Parameters
  ----------
  X : pandas.DataFrame
      Data frame containing X variables
  y : pandas.DataFrame
      Data frame containing y variables
  """
  train = X_train
  train['Direction'] = y_train['Direction'] # combine X and y
  return train

def address_class_imbalance(df, minority_classification, majority_classification,
                            minority_target_samples, majority_target_samples):
  """ Method to remove class imbalance within a data set for binary classification

  Parameters
  ----------
  df : pandas.DataFrame
      Data frame to check for and remove class imbalances from
  minority_classification : int
      the minority class
  majority_classification : int
      the majority class
  minority_target_samples: int
      number of samples to upsample to
  majority_target_samples: int
      number of samples to downsample to
  """
  # Split data frame in 2, 1 for each class
  minority_class = df[df.Direction == minority_classification]
  majority_class = df[df.Direction == majority_classification]
 
  # down sample majority and up sample minority the desired amount using helper
  # methods
  downsampled_majority = downsample_majority(majority_class, majority_target_samples)
  upsampled_minority = upsample_minority(minority_class, minority_target_samples)
  
  # combine newly sampled data frames
  new_df = upsampled_minority
  new_df = new_df.append(downsampled_majority, ignore_index=True)

  return new_df

# Upsample helper method
def upsample_minority(minority_class, target_samples):
  """ Method to up sample the minority class

  Parameters
  ----------
  minority_class : int
      the minority class
  target_samples: int
      number of samples to upsample to
  """
  return resample(minority_class, replace = True, n_samples = target_samples)

# Downsample helper method
def downsample_majority(majority_class, target_samples):
  """ Method to down sample the majority class

  Parameters
  ----------
  majority_class : int
      the majority class
  majority_target_samples: int
      number of samples to downsample to
  """
  return resample(majority_class, replace = False, n_samples = target_samples)

def find_y_mid_point(df):
  """ Method to find the mid-point between to class counts

  Parameters
  ----------
  df : pandas.DataFrame
      dataframe to find midpoint in between 2 unbalanced classes
  """
  # count number of rows belonging to each binary class
  count_class_0, count_class_1 = df.Direction.value_counts()
  # calculate mid-point
  mid_point = int(count_class_1 + ((count_class_0 - count_class_1)/2))
  return mid_point

# rebalance classes using mis point (up and down sample classes equally)
train = recombine(X_train, y_train) # recombine training sets
target_sample = find_y_mid_point(train) # determine target sample value
# up and down sample the classes to remove imbalance
train = address_class_imbalance(train, 1, 0, target_sample, target_sample)
train = shuffle_df(train) # shuffle the new training set
# Split training set back into X and y
X_train = train.iloc[:, 0:-1]
y_train = train[['Direction']]

minority_class = train[train.Direction == 1]
majority_class = train[train.Direction == 0]
print(len(minority_class), len(majority_class))

#SVM

without sentiment or finance

# Titles + content
# SVM_classifier = svm.SVC(kernel='sigmoid')
SVM_classifier = svm.LinearSVC()

#Train the model using the training sets
SVM_classifier.fit(X_train, y_train)

#Predict the response for test dataset
y_predSVM = SVM_classifier.predict(X_dev)

print(classification_report(y_dev, y_predSVM))
print(confusion_matrix(y_dev, y_predSVM))

# Titles + content
# SVM_classifier = svm.SVC(kernel='sigmoid')
SVM_classifier = svm.SVC()

#Train the model using the training sets
SVM_classifier.fit(X_train, y_train)

#Predict the response for test dataset
y_predSVM = SVM_classifier.predict(X_dev)

print(classification_report(y_dev, y_predSVM))
print(confusion_matrix(y_dev, y_predSVM))

# Titles only
# SVM_classifier = svm.SVC(kernel='sigmoid')
SVM_classifier = svm.LinearSVC()

#Train the model using the training sets
SVM_classifier.fit(X_train, y_train)

#Predict the response for test dataset
y_predSVM = SVM_classifier.predict(X_dev)

print(classification_report(y_dev, y_predSVM))
print(confusion_matrix(y_dev, y_predSVM))

without finance data but with sentiment

# Titles + content
# SVM_classifier = svm.SVC(kernel='sigmoid')
SVM_classifier = svm.LinearSVC()

#Train the model using the training sets
SVM_classifier.fit(X_train, y_train)

#Predict the response for test dataset
y_predSVM = SVM_classifier.predict(X_dev)

print(classification_report(y_dev, y_predSVM))
print(confusion_matrix(y_dev, y_predSVM))

# Titles only
# SVM_classifier = svm.SVC(kernel='sigmoid')
SVM_classifier = svm.LinearSVC()

#Train the model using the training sets
SVM_classifier.fit(X_train, y_train)

#Predict the response for test dataset
y_predSVM = SVM_classifier.predict(X_dev)

print(classification_report(y_dev, y_predSVM))
print(confusion_matrix(y_dev, y_predSVM))

with finance data and sentiment

# Content + titles
# SVM_classifier = svm.SVC(kernel='sigmoid')
SVM_classifier = svm.LinearSVC()

#Train the model using the training sets
SVM_classifier.fit(X_train, y_train)

#Predict the response for test dataset
y_predSVM = SVM_classifier.predict(X_dev)

print(classification_report(y_dev, y_predSVM))
print(confusion_matrix(y_dev, y_predSVM))

# Titles only
# SVM_classifier = svm.SVC(kernel='sigmoid')
SVM_classifier = svm.LinearSVC()

#Train the model using the training sets
SVM_classifier.fit(X_train, y_train)

#Predict the response for test dataset
y_predSVM = SVM_classifier.predict(X_dev)

print(classification_report(y_dev, y_predSVM))
print(confusion_matrix(y_dev, y_predSVM))

#RF

without finance or sentiment

# Titles + Content
rndm_fc = RandomForestClassifier()
rndm_fc = rndm_fc.fit(X_train, y_train)
y_predRF = rndm_fc.predict(X_dev)

print(classification_report(y_dev, y_predRF))
print(confusion_matrix(y_dev, y_predRF))

# Titles only
rndm_fc = RandomForestClassifier()
rndm_fc = rndm_fc.fit(X_train, y_train)
y_predRF = rndm_fc.predict(X_dev)

print(classification_report(y_dev, y_predRF))
print(confusion_matrix(y_dev, y_predRF))

without finance but with sentiment

# Titles + Content
rndm_fc = RandomForestClassifier()
rndm_fc = rndm_fc.fit(X_train, y_train)
y_predRF = rndm_fc.predict(X_dev)

print(classification_report(y_dev, y_predRF))
print(confusion_matrix(y_dev, y_predRF))

# Titles only
rndm_fc = RandomForestClassifier()
rndm_fc = rndm_fc.fit(X_train, y_train)
y_predRF = rndm_fc.predict(X_dev)

print(classification_report(y_dev, y_predRF))
print(confusion_matrix(y_dev, y_predRF))

with finance data and sentiment

# Titles + content
rndm_fc = RandomForestClassifier()
rndm_fc = rndm_fc.fit(X_train, y_train)
y_predRF = rndm_fc.predict(X_dev)

print(classification_report(y_dev, y_predRF))
print(confusion_matrix(y_dev, y_predRF))

# Titles only
rndm_fc = RandomForestClassifier()
rndm_fc = rndm_fc.fit(X_train, y_train)
y_predRF = rndm_fc.predict(X_dev)

print(classification_report(y_dev, y_predRF))
print(confusion_matrix(y_dev, y_predRF))

#GBM

without finance or sentiment

# Titles + content
GBM = GradientBoostingClassifier()
GBM = GBM.fit(X_train, y_train)
y_predGBM = GBM.predict(X_dev)

print(classification_report(y_dev, y_predGBM))
print(confusion_matrix(y_dev, y_predGBM))

#Titles only
GBM = GradientBoostingClassifier()
GBM = GBM.fit(X_train, y_train)
y_predGBM = GBM.predict(X_dev)

print(classification_report(y_dev, y_predGBM))
print(confusion_matrix(y_dev, y_predGBM))

without finance but with sentiment

# Titles + content
GBM = GradientBoostingClassifier()
GBM = GBM.fit(X_train, y_train)
y_predGBM = GBM.predict(X_dev)

print(classification_report(y_dev, y_predGBM))
print(confusion_matrix(y_dev, y_predGBM))

# Titles only
GBM = GradientBoostingClassifier()
GBM = GBM.fit(X_train, y_train)
y_predGBM = GBM.predict(X_dev)

print(classification_report(y_dev, y_predGBM))
print(confusion_matrix(y_dev, y_predGBM))

with finance and sentiment

# Titles + content
GBM = GradientBoostingClassifier()
GBM = GBM.fit(X_train, y_train)
y_predGBM = GBM.predict(X_dev)

print(classification_report(y_dev, y_predGBM))
print(confusion_matrix(y_dev, y_predGBM))

# Titles only
GBM = GradientBoostingClassifier()
GBM = GBM.fit(X_train, y_train)
y_predGBM = GBM.predict(X_dev)

print(classification_report(y_dev, y_predGBM))
print(confusion_matrix(y_dev, y_predGBM))

#Neural Network Prep

# define documents:
def create_xtext(x_df, col):
  """ Method to format data for the nerual network

  Parameters
  ----------
  X_df : pandas.DataFrame
      X values
  col:  str
      relevant columns
  """
  arr = []
  for index, row in x_df.iterrows():
    arr.append(x_df[col].loc[index]) # add text data to array row by row
  return np.array(arr)
  
# create list of concatenated vectors
def create_xtext_concat(x_df):
  """ Method to format data for the nerual network

  Parameters
  ----------
  X_df : pandas.DataFrame
      X values
  col:  str
      relevant columns
  """
  arr = []
  for index, row in x_df.iterrows():
    # concatenate title and content vectors on each row
    concatenated_vector = np.concatenate((x_df['Title'].loc[index], x_df['Content'].loc[index]),
                                         axis=None)
    arr.append(concatenated_vector) # add text data to array row by row
  return np.array(arr)

# transform titles and content into array format
train_titles = create_xtext(X_train, 'Title')
train_content = create_xtext(X_train, 'Content')

dev_titles = create_xtext(X_dev, 'Title')
dev_content = create_xtext(X_dev, 'Content')

test_titles = create_xtext(X_test, 'Title')
test_content = create_xtext(X_test, 'Content')

# prepare y variables
train_labels = y_train['Direction'].tolist()
dev_labels = y_dev['Direction'].tolist()
test_labels = y_test['Direction'].tolist()

# convert y variables into one hot vector encodings for categorical crossentropy
y_train__binary_matrix = to_categorical(y_train['Direction'], num_classes = 2)
y_dev__binary_matrix = to_categorical(y_dev['Direction'], num_classes = 2)
y_test__binary_matrix = to_categorical(y_test['Direction'], num_classes = 2)

##Concatenate variables for sequential MLP model version

# transform titles and content into array format
training = create_xtext_concat(X_train)

development = create_xtext_concat(X_dev)

testing = create_xtext_concat(X_test)

# prepare y variables
train_labels = y_train['Direction'].tolist()
dev_labels = y_dev['Direction'].tolist()
test_labels = y_test['Direction'].tolist()

train_titles.shape

#MLP

##Version1

nb_classes = 2

# Inputs
titles_input = Input(shape=(300,), name='titles_input')
content_input = Input(shape=(300,), name='content_input')
 
# the first branch operates on the first input
titles = Dense(2000, activation="relu")(titles_input)
titles = Dropout(0.5)(titles)
titles = Dense(1280, activation="relu")(titles)
titles = Model(inputs=titles_input, outputs=titles)
 
# the second branch opreates on the second input
content = Dense(2000, activation="relu")(content_input)
content = Dropout(0.5)(content)
content = Dense(1280, activation="relu")(content)
content = Model(inputs=content_input, outputs=content)
 
# combine the output of the two branches
concat = concatenate([titles.output, content.output])
 
# apply a FC layer and then a regression prediction on the
# combined outputs
fc = Dense(1000, activation="relu")(concat)
fc = Dropout(0.5)(fc)
fc = Dense(500, activation="relu")(fc)
fc = Dropout(0.1)(fc)
fc = Dense(2, activation="sigmoid")(fc)
 
# our model will accept the inputs of the two branches and
# then output a single value
model = Model(inputs=[titles.input, content.input], outputs=fc)

print(model.summary())

opt = Adam(lr=0.0001, clipnorm=1., decay=1e-6, amsgrad=True) ## decay=1e-6, 
model.compile(optimizer=opt, loss='binary_crossentropy', metrics=['accuracy'])

model.fit([train_titles, train_content], y_train__binary_matrix, batch_size = 5, epochs=15, verbose=1, validation_split=0.15)

print(history.history.keys())
# summarize history for accuracy
plt.plot(history.history['acc'])
plt.plot(history.history['val_acc'])
plt.title('model accuracy')
plt.ylabel('accuracy')
plt.xlabel('epoch')
plt.legend(['train', 'validation'], loc='upper left')
plt.show()

print("Generating test predictions...")
y_predMLP = model.predict([test_titles, test_content], verbose=1)
y_pred_bool = np.argmax(y_predMLP, axis=1)
print(classification_report(y_test, y_pred_bool))
print(confusion_matrix(y_test, y_pred_bool))

##Version 2

nb_classes = 2

# Deep Dumb MLP (DDMLP):
model = Sequential() # creates linear DNN model which consists of a linear
                     # stack of layers
# first number value is the fixed batch size of the layer
model.add(Dense(1000, input_shape=(600,))) # 2D layer specifies input shape
model.add(Activation('relu')) # activation function of layer
# Applies dropout to input
# randomly setting a fraction/rate of input units to 0 at each update during
# training, which helps prevent overfitting
model.add(Dropout(0.4))
model.add(Dense(500))
model.add(Activation('relu'))
model.add(Dropout(0.4))
model.add(Dense(nb_classes)) # final binary classification layer
model.add(Activation('softmax'))

# configure learning process
# Contains optimizer function, loss function and a list of metrics
print(model.summary())

opt = Adam(lr=0.0001, clipnorm=1., decay=1e-6, amsgrad=True) ## decay=1e-6, 
model.compile(optimizer=opt, loss='binary_crossentropy', metrics=['accuracy'])

print("training...")
history = model.fit(training, y_train__binary_matrix, batch_size = 12, epochs = 20, verbose=1, validation_split=0.15)

print(history.history.keys())
# summarize history for accuracy
plt.plot(history.history['acc'])
plt.plot(history.history['val_acc'])
plt.title('model accuracy')
plt.ylabel('accuracy')
plt.xlabel('epoch')
plt.legend(['train', 'validation'], loc='upper left')
plt.show()

y_predMLP2 = model.predict(testing, verbose=1)
y_pred_bool2 = np.argmax(y_predMLP2, axis=1)

acc = accuracy_score(test_labels, preds)
print('Accuracy: %' acc)
print(classification_report(test_labels, y_pred_bool2))
print(confusion_matrix(test_labels, y_pred_bool2))
